pyproject.toml
[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"

[project]
name = "pixi_sync_environment"
authors = [
    { name = "Bernardo Veronese", email = "bernardopveronese@gmail.com" },
]
description = "Pre-commit hook to sync a pixi environment with a traditional conda environment.yml"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.10"
keywords = []
license = { file = "LICENSE" }
classifiers = ["Programming Language :: Python :: 3"]
version = "0.1.0"
dependencies = [
    "pyyaml>=6.0.2",
]

[project.urls]
Homepage = "https://github.com/binado/pixi-sync-environment"

[project.scripts]
pixi_sync_environment = "pixi_sync_environment:main"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
lint = [
    "ruff>=0.12.7",
]
